Moved contents of oc to ocs module
