+++ tox.ini
@@ -13,7 +13,6 @@ commands = py.test \
     --ignore=tests \
     -c pytest_unittests.ini \
     --cov=mita \
-    --cov=oc \
     --cov=ocs \
     --cov=ocsci \
     --cov=resources \
