+++ utility/utils.py
@@ -736,7 +736,7 @@ def run_async(command):
 
 
 def is_cluster_running(cluster_path):
-    from oc.openshift_ops import OCP
+    from ocs.openshift_ops import OCP
     return config.RUN['cli_params'].get('cluster_path') and OCP.set_kubeconfig(
         os.path.join(cluster_path, config.RUN.get('kubeconfig_location'))
     )
